% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvega.R
\name{as_vegaspec.gg}
\alias{as_vegaspec.gg}
\title{Translate a ggplot2 object to a Vega-Lite spec}
\usage{
\method{as_vegaspec}{gg}(spec, ...)
}
\arguments{
\item{spec}{An object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{...}{Other arguments (attempt to future-proof)}
}
\value{
An object with S3 class \code{vegaspec}
}
\description{
Translate a ggplot2 object to a Vega-Lite spec
}
