% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{case_type_vl}
\alias{case_type_vl}
\title{Determine variable type}
\usage{
case_type_vl(type)
}
\arguments{
\item{type}{\code{character} (scalar) R variable-type}
}
\value{
\code{character} Vega-Lite variable-type
}
\description{
Determine variable type
}
