% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layers.R
\name{compare_data}
\alias{compare_data}
\title{Compare layer data to plot data}
\usage{
compare_data(layer_data, int_data)
}
\arguments{
\item{layer_data}{The data from a single ggplot2 layer object.}

\item{int_data}{An intermediate-form for the data created by \code{data_int()}}
}
\value{
\code{list} containing name of dataset to be used in that layer
}
\description{
Compare layer data to plot data
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))
dat <-  data_int(p$data, p$layers)
compare_data(p$layers[[1]]$data, dat)
}
