% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{create_meta_levels}
\alias{create_meta_levels}
\title{Create metadata}
\usage{
create_meta_levels(dat)
}
\arguments{
\item{dat}{Column of data for which metadata should be created.}
}
\description{
Create metadata
}
