% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{data_int}
\alias{data_int}
\title{Extract data (intermediate step)}
\usage{
data_int(data_plt, layers_plt)
}
\arguments{
\item{data_plt}{A \code{list} of data from the ggplot object.}

\item{layers_plt}{A \code{list} of layers from the ggplot object.}
}
\value{
named \code{list}, each of which has elements:
\code{metadata}, \code{variables}, \code{hash}
}
\description{
Creates an intermediate-form for the data.
}
\details{
This returns a named list of dataset objects, each object has elements:
\itemize{
\item \code{metadata}: named \code{list}, one element per variable
\item \code{variables}: the \code{data.frame} itself
\item \code{hash}: hash of \code{variables}
}
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))

data_int_demo <- data_int(p$data, p$layers)

}
\keyword{internal}
