% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{data_spc}
\alias{data_spc}
\title{Transform data from intermediate form to final form}
\usage{
data_spc(int_data)
}
\arguments{
\item{int_data}{An intermediate-form for the data created by \code{data_int()}}
}
\description{
Transform data from intermediate form to final form
}
