% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dev_example_names}
\alias{dev_example_names}
\alias{dev_example_path}
\alias{dev_example}
\alias{dev_translation}
\title{Use development examples}
\usage{
dev_example_names(type = c("ggplot", "ggspec", "vegaspec"))

dev_example_path(example = NULL, type = c("ggplot", "ggspec", "vegaspec"))

dev_example(example = NULL, type = c("ggplot", "ggspec", "vegaspec"))

dev_translation(example = NULL, type = c("ggspec", "vegaspec"))
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
\describe{
\item{\code{dev_example_names()}}{\code{character}, names of examples}
\item{\code{dev_example_path()}}{S3 object with class \code{fs_path}
(wraps \code{character}), path to R file identified by \code{example}}
\item{\code{dev_example()}}{not type-stable, depending on \code{type}, returns
either a ggplot object, a ggspec, or a vegaspec}
}
}
\description{
Note that for \code{dev_translation()}, \code{type} refers to the type \emph{to} which
the translation is made.
}
\keyword{internal}
