% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_example_shiny.R
\name{dev_example_shiny}
\alias{dev_example_shiny}
\title{Compare development-examples}
\usage{
dev_example_shiny(example = NULL, type = c("ggspec", "vegaspec"))
}
\description{
Compare development-examples
}
\keyword{internal}
