% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_gallery.R
\name{dev_gg_codeblock}
\alias{dev_gg_codeblock}
\alias{dev_gg_gallery}
\title{Create a markdown code-block from a ggplot2 example}
\usage{
dev_gg_codeblock(example)

dev_gg_gallery(example)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
\code{glue::glue()} object
}
\description{
Create a markdown code-block from a ggplot2 example
}
\keyword{internal}
