% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dev.R
\name{example_path_dev}
\alias{example_path_dev}
\alias{example_names_dev}
\alias{example_dir_dev}
\alias{gg_example_dev}
\title{Get path to example-file from development}
\usage{
example_path_dev(example = NULL, type = c("ggplot2", "ggspec",
  "vega-lite"))

example_names_dev(type = c("ggplot2", "ggspec", "vega-lite"))

example_dir_dev(type = c("ggplot2", "ggspec", "vega-lite"))

gg_example_dev(example)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}

\item{type}{\code{character}, indicates which directory to use.
Default \code{"ggplot2"}}
}
\value{
TODO
}
\description{
Get path to example-file from development
}
\keyword{internal}
