% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-example.R
\name{flip}
\alias{flip}
\title{De-factor and transpose data frame}
\usage{
flip(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{list}
}
\description{
This is a bit duplicative - it's systactic sugar to
help with the galleries.
}
\keyword{internal}
