% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{format_data_int}
\alias{format_data_int}
\title{Format the data}
\usage{
format_data_int(dat)
}
\arguments{
\item{dat}{\code{data.frame} to be formatted.}
}
\value{
\code{list} with elements \code{metadata}, \code{variables}, \code{hash}
}
\description{
Format the data
}
