% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{format_data_spec}
\alias{format_data_spec}
\title{Format data into final data form}
\usage{
format_data_spec(dat)
}
\arguments{
\item{dat}{one data frame from \code{data_int}}
}
\description{
Format data into final data form
}
