% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layers.R
\name{get_layers}
\alias{get_layers}
\title{Get layer information}
\usage{
get_layers(layer, int_data)
}
\arguments{
\item{layer}{A single ggplot2 layer object.}

\item{int_data}{An intermediate-form for the data created by \code{data_int()}.}
}
\value{
\code{list} of layer components
}
\description{
Extract necessary information from a single ggplot layer list element. Is called by \code{layer_spc()}.`
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))
dat <-  data_int(p$data, p$layers)
get_layers(p$layers[[1]], dat)
}
