% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layers.R
\name{get_mappings}
\alias{get_mappings}
\title{Extract layer mapping information}
\usage{
get_mappings(aes)
}
\arguments{
\item{aes}{An element from the list of mappings from a single ggplot2 layer object.}
}
\value{
\code{list} of mapping specifications
}
\description{
Extract the field name for the aesthetic
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))
get_mappings(p$layers[[1]]$mapping[[1]])
}
