% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scales.R
\name{get_scales}
\alias{get_scales}
\title{Get scales}
\usage{
get_scales(scale)
}
\arguments{
\item{scale}{A single ggplot2 scale object.}
}
\value{
A \code{list} of scale specifications for scales that differ from the default scales.
}
\description{
Extract necessary scales information from a single scales list. Is called by \code{scales_spc()}.`
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))
p <- p + scale_y_log10()
get_scales(p$scales$scales[[1]])
}
