% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvega.R
\name{gg2spec}
\alias{gg2spec}
\title{Translate ggplot2 object into a ggspec}
\usage{
gg2spec(plt)
}
\arguments{
\item{plt}{A ggplot2 object.}
}
\value{
\code{list}, a ggspec list.
}
\description{
Translate ggplot2 object into a ggspec
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))
gg2spec(p)
}
