% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{gg_dev_codeblock}
\alias{gg_dev_codeblock}
\title{Create a markdown code-block from a ggplot2 example}
\usage{
gg_dev_codeblock(example)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
\code{glue::glue()} object
}
\description{
Create a markdown code-block from a ggplot2 example
}
\keyword{internal}
