% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gg_example}
\alias{gg_example}
\title{Run ggplot2 example}
\usage{
gg_example(example = NULL)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
S3 object with classes \code{gg} and \code{ggplot}; a ggplot2 object
}
\description{
Use this function to run a ggplot2 example.
}
\examples{
  library("ggplot2")
  gg_example("scatterplot-iris")
}
