% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gg_example_dev}
\alias{gg_example_dev}
\title{Run ggplot2 example}
\usage{
gg_example_dev(example = NULL)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
S3 object with classes \code{gg} and \code{ggplot}; a ggplot2 object
}
\description{
This is a developer-facing version of \code{\link[=gg_example]{gg_example()}};
it can be useful if you are testing or writing out
visual regressions.
}
\keyword{internal}
