% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gg_example_file}
\alias{gg_example_file}
\title{Get path to ggplot example}
\usage{
gg_example_file(example = NULL)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
S3 object with class \code{fs_path} (wraps \code{character}),
path to R file identified by \code{example}
}
\description{
This function is useful to inspect the code used to
create the ggplto2 examples.
}
\examples{
  gg_example_file("scatterplot-iris")
}
