% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gg_example_file_dev}
\alias{gg_example_file_dev}
\title{Get path to ggplot example}
\usage{
gg_example_file_dev(example = NULL)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
S3 object with class \code{fs_path} (wraps \code{character}),
path to R file identified by \code{example}
}
\description{
This function can be useful if you are developing this package, e.g.
testing or writing out the visual regressions.
}
\keyword{internal}
