% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gg_example_name_dev}
\alias{gg_example_name_dev}
\title{Get names of ggplot examples}
\usage{
gg_example_name_dev()
}
\value{
\code{character}, names of examples
}
\description{
This function can be useful if you are developing this package, e.g.
testing or writing out the visual regressions.
}
\keyword{internal}
