% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gg_example_path}
\alias{gg_example_path}
\title{Use ggplot examples}
\usage{
gg_example_path(example = NULL)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
S3 object with class \code{fs_path} (wraps \code{character}),
path to R file identified by \code{example}
}
\description{
Use ggplot examples
}
\keyword{internal}
