% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{ggv_dev}
\alias{ggv_dev}
\alias{ggv_dev_names}
\alias{ggv_dev_path}
\title{Use development examples}
\usage{
ggv_dev(example = NULL, type = c("ggplot", "ggspec", "vegaspec"))

ggv_dev_names(type = c("ggplot", "ggspec", "vegaspec"))

ggv_dev_path(example = NULL, type = c("ggplot", "ggspec", "vegaspec"))
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
\describe{
\item{\code{ggv_dev_names()}}{\code{character}, names of examples}
\item{\code{ggv_dev_path()}}{S3 object with class \code{fs_path}
(wraps \code{character}), path to R file identified by \code{example}}
\item{\code{ggv_dev()}}{not type-stable, depending on \code{type}, returns
either a ggplot object, a ggspec, or a vegaspec}
}
}
\description{
Note that for \code{dev_translation()}, \code{type} refers to the type \emph{to} which
the translation is made.
}
\keyword{internal}
