% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{ggv_dev_display}
\alias{ggv_dev_display}
\alias{ggv_example_display}
\title{Display examples}
\usage{
ggv_dev_display(example, arrange = c("side", "top"), ...)

ggv_example_display(example, arrange = c("side", "top"),
  file_root = NULL, ...)
}
\arguments{
\item{example}{\code{character}, uses
\code{ggv_example()} to retrieve a ggplot2 object.
For \code{ggv_example_display()}, this can also be a ggplot2 object}

\item{arrange}{\code{character}, arrangement of plots}

\item{...}{Arguments passed on to internal comprison function}
}
\value{
Invisible \code{NULL}, called for side-effects
}
\description{
Display examples
}
\keyword{internal}
