% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{ggv_example}
\alias{ggv_example}
\alias{ggv_example_names}
\alias{ggv_example_path}
\title{Use ggplot examples}
\usage{
ggv_example(example = NULL)

ggv_example_names()

ggv_example_path(example = NULL)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}
}
\value{
\describe{
\item{\code{ggv_example()}}{S3 object with classes \code{gg} and \code{ggplot};
a ggplot2 object, identified by \code{example}}
\item{\code{ggv_example_names()}}{\code{character}, names of examples}
\item{\code{ggv_example_path()}}{S3 object with class \code{fs_path}
(wraps \code{character}), path to R file identified by \code{example}}
}
}
\description{
This package has a series of example files that hightlight the capabilities
of ggvega. As more capabilities are developed, more examples will be added.
}
\examples{
  library("ggplot2")
  ggv_example_names()
  ggv_example_path("scatterplot-iris")
  ggv_example("scatterplot-iris")
  ggv_example("scatterplot-iris") \%>\% as_vegaspec()
}
