% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvega.R
\name{ggvega}
\alias{ggvega}
\title{Translate ggspec object to vegaspec}
\usage{
ggvega(ggspec)
}
\arguments{
\item{ggspec}{\code{list}, a ggspec object}
}
\value{
An object with S3 class \code{vegaspec}
}
\description{
Translate ggspec object to vegaspec
}
\keyword{internal}
