% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mapping.R
\name{mapping_spc}
\alias{mapping_spc}
\title{Extract mappings}
\usage{
mapping_spc(mapping_plt)
}
\arguments{
\item{mapping_plt}{A \code{list} of defined aesthetics from the ggplot object.}
}
\value{
\code{list} of encodings with their field descriptions
}
\description{
Extract necessary information from the ggplot mappings.
}
\examples{
library(ggplot2)
p <- ggplot(data = iris, aes(x = Petal.Width, y = Petal.Length))
p <- p + geom_point()
mapping_spc(p$mapping)
}
