% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_vega.R
\name{render_vega}
\alias{render_vega}
\title{Render vega visualization
Note: this is a test function and not intended for users.}
\usage{
render_vega(x)
}
\arguments{
\item{x}{an HTML document containing an embedded JSON object}
}
\description{
Render vega visualization
Note: this is a test function and not intended for users.
}
