% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scales.R
\name{scale_spc}
\alias{scale_spc}
\title{Extract scales}
\usage{
scale_spc(scale_plt)
}
\arguments{
\item{scale_plt}{A \code{list} of scales from the ggplot object.}
}
\value{
A \code{list} of scales that differ from the default scales.
}
\description{
Extract scales
}
\examples{
library(ggplot2)
p <- ggplot(data = iris)
p <- p + geom_point(aes(x = Petal.Width, y = Petal.Length))
scale_spc(p$scales$scales)
}
