% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dev.R
\name{truncate_data_ggspec}
\alias{truncate_data_ggspec}
\title{Truncate data in ggspec}
\usage{
truncate_data_ggspec(ggspec)
}
\arguments{
\item{ggspec, }{ggspec (does this have an S3 class?)}
}
\value{
modified copy of \code{ggspec}
}
\description{
Truncate data in ggspec
}
\keyword{internal}
