% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-example.R
\name{truncate_data_vegaspec}
\alias{truncate_data_vegaspec}
\title{Truncate data in vegaspec}
\usage{
truncate_data_vegaspec(vegaspec)
}
\arguments{
\item{vegaspec, }{vegaspec (does this have an S3 class?)}
}
\value{
modified copy of \code{vegaspec}
}
\description{
Truncate data in vegaspec
}
\keyword{internal}
