% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dev.R
\name{example_path_dev}
\alias{example_path_dev}
\alias{example_names_dev}
\alias{example_dir_dev}
\title{Get path to example-file from development}
\usage{
example_path_dev(example = NULL, type = c("ggplot2", "ggspec",
  "vega-lite"))

example_names_dev(type = c("ggplot2", "ggspec", "vega-lite"))

example_dir_dev(type = c("ggplot2", "ggspec", "vega-lite"))
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}

\item{type}{\code{character}, indicates which directory to use.
Default \code{"ggplot2"}}
}
\value{
\describe{
\item{\code{gg_example_name()}}{\code{character}, names of examples}
\item{\code{gg_example_path()}}{S3 object with class \code{fs_path}
(wraps \code{character}), path to R file identified by \code{example}}
\item{\code{gg_example()}}{S3 object with classes \code{gg} and \code{ggplot};
a ggplot2 object, identified by \code{example}}
}
}
\description{
Get path to example-file from development
}
\keyword{internal}
