% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{ggv_dev_display}
\alias{ggv_dev_display}
\alias{ggv_example_display}
\title{Display examples}
\usage{
ggv_dev_display(example, arrange = c("side", "top"), ...)

ggv_example_display(example, arrange = c("side", "top"),
  vl_width = 275, vl_height = 275, gg_width = 400, gg_height = 320)
}
\arguments{
\item{example}{\code{character} or \code{ggplot2} object, if \code{character}, uses
\code{ggv_example()} to retrieve a ggplot2 object.}

\item{arrange}{\code{character}, arrangement of plots}

\item{vl_width}{\code{integer}, width of VL chart (px.)}

\item{vl_height}{\code{integer}, height of VL chart (px.)}

\item{gg_width}{\code{integer}, height of GG chart (px.)}

\item{gg_height}{\code{integer}, width of GG chart (px.)}
}
\value{
Invisible \code{NULL}, called for side-effects
}
\description{
Display examples
}
\keyword{internal}
