% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_gallery.R
\name{ggv_dev_display}
\alias{ggv_dev_display}
\title{Create a markdown code-block from a ggplot2 example}
\usage{
ggv_dev_display(example, arrange = c("side", "top"), vl_width = 275,
  vl_height = 275, gg_width = 400, gg_height = 320)
}
\arguments{
\item{example}{\code{character}, name of the example,
\code{NULL} prints message containing names of all examples}

\item{arrange}{\code{character}, arrangement of plots}

\item{vl_width}{\code{integer}, width of VL chart (px.)}

\item{vl_height}{\code{integer}, height of VL chart (px.)}

\item{gg_width}{\code{integer}, height of GG chart (px.)}

\item{gg_height}{\code{integer}, width of GG chart (px.)}
}
\value{
\code{glue::glue()} object
}
\description{
Create a markdown code-block from a ggplot2 example
}
\keyword{internal}
