% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvega.R
\name{ggvega}
\alias{ggvega}
\title{Translate a ggspec}
\usage{
ggvega(ggspec)
}
\arguments{
\item{ggspec}{\code{list}, ggspec object}
}
\value{
An object with S3 class \code{vegaspec}
}
\description{
Translate a ggspec
}
