% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ggvega_examples}
\alias{ggvega_examples}
\title{Examples}
\format{An object of class \code{list} of length 9.}
\usage{
ggvega_examples
}
\description{
A list of example objects. In addition to showing the types of plots we can
translate, we use these examples in our test suite and in our documentation
to make sure things continue to work as we expect.
}
\details{
This is a named list; each name refers to an example. Each element is an R6
object with accessors:

\describe{
\item{\verb{$plot}}{returns the ggplot object.}
\item{\verb{$path}}{returns the path to the R file containing ggplot2 code.}
\item{\verb{$code}}{returns the code that builds the ggplot object.}
\item{\verb{$as_vegaspec()}}{returns the Vega-Lite translation.}
\item{\verb{$display()}}{used in documentation, works only with knitr.}
}

The naming convention is to start with the geom being used, e.g. \code{point}, \code{bar}.
You can use \code{names(ggvega_examples)} to return the names of all the examples.
Because this is a list of R6 objects, in the RStudio IDE you can use
autocomplete with the \code{$} operator.
}
\examples{
  library("ggplot2")
  # print method
  # ggvega_examples$point_iris$plot returns the ggplot object
  ggvega_examples$point_iris
  # return the ggplot code
  ggvega_examples$point_iris$code
  # translate to Vega-Lite
  ggvega_examples$point_iris$as_vegaspec()

}
\keyword{datasets}
